Added latitude and longitude fields as well as code to edit and store them in the database

+++ app/src/main/java/se/c19aky/geolocations/Location.kt
@@ -8,5 +8,5 @@ import java.util.*
 @Entity
 data class Location(@PrimaryKey val id: UUID = UUID.randomUUID(),
                     @ColumnInfo var name: String = "Location",
-                    @ColumnInfo var latitude: Int = 0,
-                    @ColumnInfo var longitude: Int = 0)
+                    @ColumnInfo var latitude: Double = 0.0,
+                    @ColumnInfo var longitude: Double = 0.0)
